+++ tests/src/operation_tests.rs
@@ -71,6 +71,7 @@ async fn test_basic_t22_operations() {
         ix_metadata: None,
         ix_scaled_ui_amount: None,
         ix_default_account_state: None,
+        ix_permissioned_burn: false,
     };
 
     initialize_mint(
@@ -262,6 +263,7 @@ async fn test_t22_extension_operations() {
         ix_metadata: None,
         ix_scaled_ui_amount: None,
         ix_default_account_state: None,
+        ix_permissioned_burn: false,
     };
 
     initialize_mint(
@@ -399,6 +401,7 @@ async fn test_t22_transfer_operations() {
         ix_metadata: None,
         ix_scaled_ui_amount: None,
         ix_default_account_state: None,
+        ix_permissioned_burn: false,
     };
 
     initialize_mint(
@@ -533,6 +536,7 @@ async fn test_p2p_transfer_direct_spl() {
         ix_metadata: None,
         ix_scaled_ui_amount: None,
         ix_default_account_state: None,
+        ix_permissioned_burn: false,
     };
 
     initialize_mint(
@@ -697,6 +701,7 @@ async fn test_transfer_hook_extra_account_metas_init_update_trim() {
         ix_metadata: None,
         ix_scaled_ui_amount: None,
         ix_default_account_state: None,
+        ix_permissioned_burn: false,
     };
 
     initialize_mint(